2020-10-16 check test machine

+++ 02_Get_Next_Line/get_next_line.h
@@ -2,6 +2,9 @@
 # define GET_NEXT_LINE_H
 # include <unistd.h>
 # include <stdlib.h>
+# include <stdio.h>
+# include <fcntl.h>
+# include <limits.h>
 
 # ifndef BUFFER_SIZE
 # define BUFFER_SIZE	10
@@ -14,7 +17,7 @@
 int		get_next_line(int fd, char **line);
 
 char	*ft_strchr(const char *s, int c);
-char	*ft_strndup(const char *src, int n);
+char	*ft_strndup(const char *src, size_t n);
 char	*ft_strjoin(char const *s1, char const *s2);
 size_t	ft_strlen(const char *s);
 void	*ft_memcpy(void *dst, const void *src, size_t n);
